### 2023-11-09 #76 @ADevillers
[Re] A Simple Framework for Contrastive Learning of Visual Representations
**Original article:** T. Chen, S. Kornblith, M. Norouzi, and G. Hinton. “A simple framework for contrastive learning of visual repre-
sentations.” In: International conference on machine learning. PMLR, 2020, pp. 1597–1607

**Scientific domain:** Representation Learning
**Suggested editor:** @rougier
