### 2023-12-04 #76 @benoit-girard
@pps121 would you like to review this paper? And possibly (or aletrnatively) its companion paper https://github.com/ReScience/submissions/issues/77 ?
Let me know!
