### 2023-12-05 #76 @charlypg
Hello everybody.
I am going to review SimCLR then BYOL. I have a lot to do for my own research during 2 weeks but I think I can do your review before the 25. Is it ok for you ?
It will also depend on the required computational resources.
