### 2023-12-21 #76 @benoit-girard
@bsciolla would you like to review this paper? And possibly (or alternatively) its companion paper https://github.com/ReScience/submissions/issues/77 ?
Let me know!
