### 2024-01-19 #76 @benoit-girard
@mo-arvan would you like to review this paper? And possibly (or alternatively) its companion paper https://github.com/ReScience/submissions/issues/77 ?
Let me know!
