### 2024-01-31 #76 @charlypg
Hello everybody. I am really sorry for the delay.
First of all, thank you for this work that may benefit to the community because reproduction in machine learning is always complicated as tips and tricks are not always precised in the articles themselves.
Here are two lists, one for the good aspects and another one for the problems I encountered.

**Good :**
- Implementation tips and tricks are precised in the article
- The code is proper and clear which benefits to the community
- I could reproduce CIFAR top-1 accuracy results
- (PS: How to reproduce top-5 ?)

**Problems :**
- Config for single GPU
   - Please provide minimum config (cuda version + conda env + requirements). I had some problems because I didn't have the right cuda/cudnn version. Even though it takes only 5-10min to fix, I think it is important.
   - "six" module is not in requirements

- For evaluation I had the following problem :
"Error tracker : world_size missing argument for tracker". So I set it to 1.
What does world_size mean and how to set it ?

- With world_size=1 I could reproduce eval results for CIFAR but not for ImageNet on Jean Zay. In the logs I obtain 59% of top-1 accuracy instead of 70% in the article and I also obtain a warning with it :
"WARNING: A reduction issue may have occurred (abs(50016.0 - 1563.0*1) >= 1)."
