### 2024-02-02 #76 @charlypg
Dear @ADevillers ,

Thank you for your response.
I will try the evaluation on other checkpoints. By the way, what do "even" and "odd" mean regarding checkpoints ?

Thank you in advance,
Charly PECQUEUX--GUÉZÉNEC
