### 2024-02-07 #76 @benoit-girard
> > @MiWeiss would you like to review this paper? And possibly (or alternatively) its companion paper #77 ? Let me know!
>
> Hi @benoit-girard. Unfortunately, I am currently not available - and I am afraid I also would not have quite the compute needed to run the code of this paper ;-)

Thanks a lot for your answer.
