### 2024-02-07 #76 @benoit-girard
@ReScience/reviewers I am looking for a reviewer with expertise in machine learning to review this submission and possibly (or alternatively) its companion paper https://github.com/ReScience/submissions/issues/77
