### 2024-02-07 #76 @charlypg
Dear @ADevillers ,

Thank you for your answer.

I have a question about the training. Once the job corresponding to "run_simclr_imagenet.slurm" has successfully ended, I only obtain one checkpoint of the form "expe_[job_id]_[epoch_number].pt".
If I understand your paper well  ("Jobs too long and checkpoints" paragraph), you submit the same slurm multiple times to reach 800 epochs ?
If yes, is the checkpoint, from which you start training, the only thing you modify in the slurm script ?

Best regards,
Charly PECQUEUX--GUÉZÉNEC
