### 2024-04-20 #76 @ADevillers
Dear @charlypg,

Your runs should normally be stored in the "runs" folder under a format readable by tensorboard and contains all the curves (including Top-5 acc).

Note that, when starting from a checkpoint, the data will append to the file corresponding to the run of the checkpoint. Therefore, a run on ImageNet, even if it requires 6 to 7 restarts from a checkpoint, will only produce one file (that will contain everything).

To find out where the issue could be, can you please answer the following questions:
1. Is your "runs" folder empty?
2. Have you been able to open tensorboard with the "runs" folder?
3. If so, do you see any runs/curves?
4. Are you able to find in the runs list the ones starting with the same ID as the first job of your run?
5. If so, is there any curve you are able to see for these runs?

Best,
Alexandre DEVILLERS
